+++ src/websocket-server/src/Exception/Handler/WebSocketExceptionHandler.php
@@ -9,7 +9,6 @@ declare(strict_types=1);
  * @contact  group@hyperf.io
  * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
  */
-
 namespace Hyperf\WebSocketServer\Exception\Handler;
 
 use Hyperf\Contract\StdoutLoggerInterface;