+++ src/di/tests/ExceptionStub/DemoInjectException.php
@@ -9,7 +9,6 @@ declare(strict_types=1);
  * @contact  group@hyperf.io
  * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
  */
-
 namespace HyperfTest\Di\ExceptionStub;
 
 use Hyperf\Di\Annotation\Inject;