+++ src/translation/src/Translator.php
@@ -9,7 +9,6 @@ declare(strict_types=1);
  * @contact  group@hyperf.io
  * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
  */
-
 namespace Hyperf\Translation;
 
 use Countable;
@@ -368,9 +367,6 @@ class Translator implements TranslatorInterface
 
     /**
      * Set the parsed value of a key.
-     *
-     * @param string $key
-     * @param array $parsed
      */
     public function setParsedKey(string $key, array $parsed)
     {
@@ -489,9 +485,6 @@ class Translator implements TranslatorInterface
 
     /**
      * Parse an array of namespaced segments.
-     *
-     * @param string $key
-     * @return array
      */
     protected function parseNamespacedSegments(string $key): array
     {