+++ src/crontab/src/Mutex/RedisServerMutex.php
@@ -52,9 +52,9 @@ class RedisServerMutex implements ServerMutex
         $result = (bool) $redis->set($mutexName, $this->macAddress, ['NX', 'EX' => $crontab->getMutexExpires()]);
 
         if ($result === true) {
-            Coroutine::create(function() use ($crontab, $redis, $mutexName) {
-               $exited = CoordinatorManager::until(Constants::WORKER_EXIT)->yield($crontab->getMutexExpires());
-               $exited && $redis->del($mutexName);
+            Coroutine::create(function () use ($crontab, $redis, $mutexName) {
+                $exited = CoordinatorManager::until(Constants::WORKER_EXIT)->yield($crontab->getMutexExpires());
+                $exited && $redis->del($mutexName);
             });
             return $result;
         }