+++ src/grpc-server/src/CoreMiddleware.php
@@ -130,7 +130,7 @@ class CoreMiddleware extends HttpCoreMiddleware
                         $ref = $definition['ref'];
                         $class = ReflectionManager::reflectClass($ref);
                         $parentClass = $class->getParentClass();
-                        if ($parentClass->getName() === ProtobufMessage::class) {
+                        if ($parentClass && $parentClass->getName() === ProtobufMessage::class) {
                             $request = $this->request();
                             $stream = $request->getBody();
                             return Parser::deserializeMessage([$class->getName(), null], $stream->getContents());