+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - [#1563](https://github.com/hyperf/hyperf/pull/1563) Fixed crontab's `onOneServer` option not resetting mutex on shutdown.
 - [#1565](https://github.com/hyperf/hyperf/pull/1565) Reset transaction level to zero, when reconnent to mysql server.
+- [#1579](https://github.com/hyperf/hyperf/pull/1579) Fixed refresh command's `step` is int.
 
 # v1.1.25 - 2020-04-09