+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ## Changed
 
 - [#1560](https://github.com/hyperf/hyperf/pull/1560) Changed functions of file to `filesystem` for `FileSystemDriver` in `hyperf/cache`.
+- [#1568](https://github.com/hyperf/hyperf/pull/1568) Changed `\Redis` to `RedisProxy` for `RedisDriver` in `async-queue`.
 
 # v1.1.25 - 2020-04-09