+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [#1565](https://github.com/hyperf/hyperf/pull/1565) Reset transaction level to zero, when reconnent to mysql server.
 - [#1572](https://github.com/hyperf/hyperf/pull/1572) Fixed parent class does not exists in `Hyperf\GrpcServer\CoreMiddleware`.
 - [#1577](https://github.com/hyperf/hyperf/pull/1577) Fixed `describe:routes` command's `server` option not take effect
+- [#1579](https://github.com/hyperf/hyperf/pull/1579) Fixed refresh command's `step` is int.
 
 ## Changed